--- packages/frontend/src/redux/selectors/mainSelectors/selectAccountState.js
@@ -1,1 +1,0 @@
-export default (state, { accountId }) => state[accountId] || {};
