--- packages/frontend/src/redux/selectors/mainSelectors/selectSliceByAccountId.js
@@ -1,5 +1,0 @@
-import { createSelector } from 'reselect';
-
-import selectAccountState from './selectAccountState';
-
-export default (sliceName, initialState) => createSelector(selectAccountState, (accountState) => accountState[sliceName] || initialState);
