+++ packages/frontend/src/redux/slices/recoveryMethods/index.js
@@ -5,7 +5,7 @@ import { createSelector } from 'reselect';
 import { wallet } from '../../../utils/wallet';
 import handleAsyncThunkStatus from '../../reducerStatus/handleAsyncThunkStatus';
 import initialStatusState from '../../reducerStatus/initialState/initialStatusState';
-import createParameterSelector from '../../selectors/mainSelectors/createParameterSelector';
+import { createParameterSelector } from '../../selectors/topLevel';
 
 const SLICE_NAME = 'recoveryMethods';
